cti_app_dj/requirements.txt
anyio==4.9.0
asgiref==3.8.1
certifi==2025.4.26
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
sniffio==1.3.1
sqlparse==0.5.3
tenacity==9.1.2
typing_extensions==4.13.2
tzdata==2025.2
